requirements.txt
pytest~=7.2.2
black~=23.1.0
fastapi~=0.110.1
uvicorn[standard]~=0.35.0
pydantic~=2.6.4
pydantic[email]
mangum~=0.17.0
boto3~=1.34.118
flask
firebase-admin~=6.5.0
geopy~=2.4.1
timezonefinder~=6.5.0
nrdtech-aws-utils~=0.0.4
python-dotenv~=1.0.1
redis~=5.0.4
snowflake-connector-python
bidict==0.23.1
msal~=1.24.1
asyncpg~=0.30.0
databases~=0.9.0
pytest-asyncio
SQLAlchemy~=2.0.30
botocore~=1.34.118
nrdtech_aws_utils~=0.0.4
requests~=2.32.3
newrelic~=9.11.0
openai>=1.97.0
python-multipart~=0.0.12
starlette~=0.37.2
pillow~=10.4.0
opensearch-py[async]~=2.8.0
pytz~=2024.2
httpx==0.27.2
numpy==1.24.4
cryptography~=43.0.3
certifi
aiohttp~=3.12.13
beautifulsoup4
